data_pusher/requirements.txt
djangorestframework
requests